apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3
pydantic>=2.9.0
redis>=5.2.0
celery>=5.4.0
pyjwt[crypto]>=2.10.1
